+++ __tests__/01_basic_spec.js
@@ -1,7 +1,7 @@
-import React, { StrictMode, Suspense } from 'react';
+import React, { Suspense } from 'react';
 import { render, cleanup } from '@testing-library/react';
 
-import { useFetch } from '../src/index';
+import { createResource, useResource } from '../src/index';
 
 const sleep = ms => new Promise(resolve => setTimeout(resolve, ms));
 
@@ -9,23 +9,21 @@ describe('basic spec', () => {
   afterEach(cleanup);
 
   it('should have a function', () => {
-    expect(useFetch).toBeDefined();
+    expect(createResource).toBeDefined();
+    expect(useResource).toBeDefined();
   });
 
   it('should create a component', async () => {
     fetch.mockResponse(JSON.stringify('test data'));
+    const resource = createResource('http://...');
     const DisplayRemoteData = () => {
-      const { error, data } = useFetch('http://...');
-      if (error) return <span>Error: {error.message}</span>;
-      if (!data) return null; // this is important
+      const { data } = useResource(resource);
       return <span>RemoteData: {data}</span>;
     };
     const App = () => (
-      <StrictMode>
-        <Suspense fallback={<span>Loading...</span>}>
-          <DisplayRemoteData />
-        </Suspense>
-      </StrictMode>
+      <Suspense fallback={<span>Loading...</span>}>
+        <DisplayRemoteData />
+      </Suspense>
     );
     const { container } = render(<App />);
     expect(container).toMatchSnapshot();